# internal/shared/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Withdraw Service API
  version: 1.0.0
  description: API untuk autentikasi, pengelolaan user, wallet, dan withdraw.
servers:
  - url: .
    description: Server tempat dokumentasi ini disajikan
...
paths:
...
  /api/transaction/withdraw:
    post:
      tags: [Transactions]
      operationId: withdraw
      summary: Melakukan withdraw dari wallet user yang login
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawRequest"
      responses:
        "200":
          description: Withdraw berhasil diproses
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/WithdrawData"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Payload atau parameter tidak valid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Token tidak tersedia atau tidak valid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Kesalahan internal server
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    SuccessResponse:
      type: object
      required: [status, message, error]
      properties:
        status:
          type: boolean
          enum: [true]
        message:
          type: string
        data:
          nullable: true
        error:
          nullable: true
          example: null
    ErrorResponse:
      type: object
      required: [status, message, error]
      properties:
        status:
          type: boolean
          enum: [false]
        message:
          type: string
        error:
          nullable: true
...
    UserSummary:
      type: object
      required: [user_id, name]
      properties:
        user_id:
          type: string
          format: uuid
        name:
          type: string
    WithdrawRequest:
      type: object
      additionalProperties: false
      required: [amount]
      properties:
        amount:
          type: integer
          format: int64
          minimum: 1
          example: 40000
        remark:
          type: string
          nullable: true
          example: ATM
    WithdrawData:
      type: object
      required: [transaction_id, reference_number, transaction, user]
      properties:
        transaction_id:
          type: string
          format: uuid
        reference_number:
          type: string
        transaction:
          type: object
          required: [amount, type, balance_now]
          properties:
            amount:
              type: integer
              format: int64
            type:
              type: string
              enum: [withdraw]
            balance_now:
              type: integer
              format: int64
            remark:
              type: string
              nullable: true
        user:
          $ref: "#/components/schemas/UserSummary"